moves get_all_repo_files to core
